requirements.txt
streamlit>=1.32
locust>=2.25
pandas>=2.0
plotly>=5.20
python-dotenv>=1.0
watchdog
reportportal-client>=5.5.0
pydantic-settings>=2.0